requirements.txt
aiogram==3.1.1
bs4
lxml
python-dotenv==0.21.0
pytz
requests==2.31.0
sqlalchemy==2.0.21
